Add warnings to error response for sklearn Pipelines (#114)
* Add warnings to error response

See #92

When calling the prediction pipeline produces warnings, it is treated as
an error. This includes situations where the model output *might* be
correct but we're not sure, most notably when there is a mismatch of the
version of the loaded sklearn model and the sklearn version of the
runtime.

Although we still return the model outputs for the user, we don't show
them what warnings were actually encountered. This makes debugging
difficult. Therefore, we want to show them the warnings.

This requires two changes. First, we need to add the encountered
warnings to the response. Second, we need to display these warnings.
This PR implements the first step.

On top of that, this PR introduces a small refactoring to reduce the
amount of code duplication (see the last paragraph of this comment:
https://github.com/huggingface/api-inference-community/pull/92#issuecomment-1231776124).
This refactoring results in no functional changes.

* Fix linting/formatting/imports

* Remove unnecessary super() call

I started with implementing the same method on parent class, requiring
super, then moved away, but forgot to remove the super call.

* Add missing word to comment

* Address reviewer comments

1. Remove "useless" delegation call.
2. Make _get_output an abstractmethod.